More refactoring for GUIfication
