+++ pyproject.toml
@@ -27,4 +27,4 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-blamscamp = 'blamscamp.__init__:main'
+blamscamp = 'blamscamp.__main__:main'
